1.5-r2 + Moved keymaps to module directory + Updated paths + Minor changes in md files
