+++ kop/ciastkove_ulohy.md
@@ -4,5 +4,5 @@
 - [x] Umožňiť zobrazenie konkrétheho typu dát na grafe
 - [x] Spraviť graf interaktívnym (zoom, zobrazovanie aktuálnych hodnôt)
 - [x] Umožňiť používateľovi uloženie grafu aj s úpravami do obrazového súboru
-- [ ] Umožňiť používateľovi úpravu dát v grafe
-- [ ] Umožňiť uloženie modifikovaných dát
+- [x] Umožňiť používateľovi úpravu dát v grafe
+- [x] Umožňiť uloženie modifikovaných dát
